% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMean.R
\name{calcMean}
\alias{calcMean}
\title{Calculate mean for duplicated samples}
\usage{
calcMean(data, duplicated = TRUE, max.dup = 2)
}
\arguments{
\item{data}{A list of data.frames, each containing at least \code{sampleID}, \code{Gene}, \code{Task}.}

\item{duplicated}{Logical; whether to process duplicated samples.}

\item{max.dup}{Integer; maximum number of duplications to consider.}
}
\value{
A list of data.frames with averaged values for duplicates.
}
\description{
This function computes means of numeric columns for duplicated \code{sampleID}s
up to \code{max.dup} and combines them with non-duplicated samples.
}
