% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRatios.R
\name{calcRatios}
\alias{calcRatios}
\title{Calculate ratios between control and genes}
\usage{
calcRatios(data)
}
\arguments{
\item{data}{A list of data.frames. The first element is the control, others are genes.
Each data.frame must have columns: \code{sampleID}, \code{Gene}, \code{CT}, \code{Quantity}.}
}
\value{
A list of data.frames, each containing the ratios for one gene.
}
\description{
This function calculates the CT and Quantity ratios (\%) of each gene
relative to the control (assumed to be the first element in \code{data}).
}
