% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_NA.R
\name{check_NA}
\alias{check_NA}
\title{Check and remove NA/unknown samples from qPCR data}
\usage{
check_NA(
  data,
  CT.name,
  Quantity.name,
  Well.name,
  sampleID,
  Detector.name,
  dir.name
)
}
\arguments{
\item{data}{A list of data.frames to process.}

\item{CT.name}{Character vector of length 1 or 2. Column name(s) for CT values.}

\item{Quantity.name}{Character. Column name for Quantity.}

\item{Well.name}{Character vector of length 1 or 2. Column name(s) for Well position.}

\item{sampleID}{Character. Column name for the sample identifier.}

\item{Detector.name}{Character vector of length 1 or 2. Column name(s) for the gene detector.}

\item{dir.name}{Character. Path to the main output directory.}
}
\value{
A list of cleaned data.frames, and writes a CSV with removed samples.\
}
\description{
This function checks a list of data.frames for missing or invalid qPCR values,
removes samples with NA or invalid \code{CT}/\code{Quantity}, and returns a list of cleaned data.
}
