% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConcat.R
\name{geneConcat}
\alias{geneConcat}
\title{Concatenate rows by gene of interest}
\usage{
geneConcat(data, genes)
}
\arguments{
\item{data}{A list of data.frames, each containing at least a \code{Gene} column.}

\item{genes}{A character vector of gene names to filter by.}
}
\value{
A list of data.frames, each corresponding to one gene in \code{genes}.
}
\description{
This function filters and concatenates rows from a list of data.frames
for each specified gene of interest.
}
