% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load and organize data files in a project}
\usage{
loadData(
  skipping.rows = 0,
  separator = ",",
  filePattern,
  xlsx = FALSE,
  csv = FALSE,
  dir.name
)
}
\arguments{
\item{skipping.rows}{Integer. Number of rows to skip when reading the files.}

\item{separator}{Character. Separator for CSV files (default \code{","}).}

\item{filePattern}{Character. Regex pattern to search for data files.}

\item{xlsx}{Logical. Whether to read \code{.xlsx} files.}

\item{csv}{Logical. Whether to read \code{.csv} files.}

\item{dir.name}{Character. Name of the main directory to create.}
}
\value{
A list of data.frames with the imported data, named after the files.
}
\description{
This function creates a directory with subfolders, moves files matching a pattern,
and loads \code{.xls} or \code{.csv} files into a list of data.frames.
}
