% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{Plot boxplots of gene quantities}
\usage{
plotBar(
  data,
  colors,
  genes,
  ylimit,
  MeanQuantity = FALSE,
  RatioQuantity = FALSE,
  dir.name = "."
)
}
\arguments{
\item{data}{A data.frame with sampleID and gene quantities.}

\item{colors}{Character vector. Colors to use for the gene fill.}

\item{genes}{Character vector. Names of the genes to include in the plot.}

\item{ylimit}{Numeric vector of length 2. Y-axis limits.}

\item{MeanQuantity}{Logical. If TRUE, plots mean quantity per gene.}

\item{RatioQuantity}{Logical. If TRUE, plots ratio quantity per gene.}

\item{dir.name}{Character. Path to the main output directory.}
}
\value{
A ggplot2 object with the boxplot.
}
\description{
This function generates boxplots (and jitter) for either mean quantities or ratio quantities
across genes. It also saves the plot as a TIFF file.
}
